chore: restructure the code and removed useEffect from page and implement dynamic import
