+++ apps/docs/app/collection/Playground/Blocks/settings/utils/SettingSwitch.tsx
